+++ src/org/minima/system/commands/base/sphincs.java
@@ -112,6 +112,7 @@ public class sphincs extends Command {
 				
 				//Create the file
 				File backupfile = MiniFile.createBaseFile(file);
+				MiniFile.validateFileAccess(backupfile);
 				if(backupfile.exists()) {
 					backupfile.delete();
 				}
@@ -135,6 +136,7 @@ public class sphincs extends Command {
 			if(existsParam("file")) {
 				String file = getParam("file");
 				File ff = MiniFile.createBaseFile(file);
+				MiniFile.validateFileAccess(ff);
 				if(!ff.exists()) {
 					throw new CommandException("File does not exist : "+ff.getAbsolutePath());
 				}
